Add movies store
