+++ src/components/SearchContent.vue
@@ -1,20 +1,14 @@
 <script lang="ts" setup>
-import { ref } from 'vue'
-import type { Movie } from '../utils/fetchMovies'
 import MovieCard from './MovieCard.vue'
-import fetchMovies from '../utils/fetchMovies'
+import { useMoviesStore, type Movie } from '@/stores/movies'
 
-const { movies } = defineProps<{
-  movies: Movie[]
-}>()
+const moviesStore = useMoviesStore()
 const emit = defineEmits<{
   (e: 'click', movie: Movie): void
 }>()
 
-// @Todo with piana
-const moviesSearchedByGenre = ref<Movie[]>()
 const searchByGenre = async (movie: Movie) => {
-  moviesSearchedByGenre.value = await fetchMovies(movie.genres[0], 'genre', 'rating')
+  moviesStore.fetchMovies(movie.genres[0], 'genre', 'release date')
   emit('click', movie)
   window.scrollTo(0, 0)
 }
@@ -22,7 +16,7 @@ const searchByGenre = async (movie: Movie) => {
 
 <template>
   <div class="bg-netflix-gray-23 px-16 py-16 flex flex-wrap gap-14">
-    <div v-for="movie in movies">
+    <div v-for="movie in moviesStore.movies">
       <MovieCard
         :key="movie.posterurl"
         :title="movie.title"
